refactor: Move determiner to a separate function
