+++ index.js
@@ -2,24 +2,15 @@
 
 //@ts-check
 
-// index.js
 import { Client, Events, GatewayIntentBits } from "discord.js";
 import { config } from "dotenv";
-import { readFile } from "fs/promises";
-
-/**
- * 用於拼寫檢查的規則項目
- * @typedef {Object} SpellingRule
- * @property {string} wrong - 錯誤的拼寫
- * @property {string} correct - 正確的拼寫
- * @property {boolean} caseSensitive - 是否區分大小寫
- */
+import { Determiner } from "./determiner";
 
 /**
  * 拼寫錯誤的資訊
  * @typedef {Object} SpellingMistake
- * @property {string} wrong - 錯誤的拼寫
- * @property {string} correct - 正確的拼寫
+ * @property {string} wrong 錯誤的拼寫
+ * @property {string} correct 正確的拼寫
  */
 
 // 載入環境變數
@@ -34,57 +25,16 @@ const client = new Client({
     ],
 });
 
-/** @type {SpellingRule[]} */
-let spellingChecks = [];
-
-/**
- * 載入拼寫檢查規則
- * @returns {Promise<void>}
- */
-async function loadRules() {
-    try {
-        const rulesData = await readFile("./rules.json", "utf-8");
-        /** @type {{rules: SpellingRule[]}} */
-        const parsedData = JSON.parse(rulesData);
-        spellingChecks = parsedData.rules;
-        console.log("成功載入規則檔案");
-    } catch (error) {
-        console.error("載入規則檔案時發生錯誤:", error);
-        process.exit(1);
-    }
-}
-
-/**
- * 檢查訊息內容中的拼寫錯誤
- * @param {string} content - 要檢查的訊息內容
- * @returns {SpellingMistake[]} - 找到的拼寫錯誤列表
- */
-function checkSpelling(content) {
-    /** @type {SpellingMistake[]} */
-    const mistakes = [];
-
-    for (const check of spellingChecks) {
-        const searchText = check.caseSensitive
-            ? content
-            : content.toLowerCase();
-        const searchWord = check.caseSensitive
-            ? check.wrong
-            : check.wrong.toLowerCase();
-
-        if (searchText.includes(searchWord)) {
-            mistakes.push({
-                wrong: check.wrong,
-                correct: check.correct,
-            });
-        }
-    }
-
-    return mistakes;
-}
+console.log("正在載入規則檔案……");
+const determiner = await Determiner.fromFile("./rules.json").catch(err => {
+    console.error(err);
+    process.exit(1);
+}).finally(() => {
+    console.log("成功載入規則檔案");
+})
 
 // 當 bot 準備就緒時
 client.once(Events.ClientReady, async () => {
-    await loadRules();
     console.log(`Logged in as ${client.user?.tag}`);
 });
 
@@ -95,7 +45,7 @@ client.on(Events.MessageCreate, async message => {
     // 忽略 bot 的訊息
     if (message.author.bot) return;
     try {
-        const mistakes = checkSpelling(message.content);
+        const mistakes = determiner.checkSpelling(message.content);
 
         if (mistakes.length > 0) {
             const response = mistakes
